requirements.txt
Django==4.0.2
django-cors-headers==3.11.0
django-environ==0.8.1
django-model-utils==4.2.0
django-neomodel==0.0.7
django-rest-framework==0.1.0
djangorestframework==3.13.1
mysqlclient==2.1.0
neo4j==5.8.0 
neobolt==1.7.17
neomodel==5.0.0
Pillow==8.4.0
pyotp==2.6.0
pytz==2021.3
boto3
django-storages
python-magic
django-ses
Faker
django-channels==0.7.0
channels-redis==3.4.1
djangorestframework-simplejwt==5.2.0